Get data from INMET website

get_inmet_data_by_year() downloads data of a specific year from the official INMET website and prepreprocess it.

Usage

get_inmet_data_by_year(
  year,
  first.day = NA,
  last.day = NA,
  vars = NULL,
  stations = NULL
)

Arguments

year

numeric; integer number between 2000 and 2024.

first.day, last.day

character in the format "mm-dd". If NA (default), the first/last day of the year is considered.

vars

character vector; variables to be collected. If NULL (default), all variables are collected.

stations

character vector; stations to be collected. If NULL (default), all stations are collected.

Value

A tibble::tibble() containing the required data. Errors may arise if:

year is not specified.

You're trying to collect data before 2000-May-07.

fist.day doesn't comes before last.day.

Just one of first.day & last.day is passed.

Your PC is not connected to the internet.

First day in record

If you're trying to collect data of the year 2000 and the first.day is before 2000-May-07, you're only getting data from 2000-May-07 and beyond because that was (allegedly) the day INMET automatic stations began operations. Note that some stations began collecting actual data some days after 2000-May-07.

Examples

if (FALSE) { # \dontrun{
# `first.day` & `last.day` must be in format mm-dd, here the code wouldn't run
# because it is in the format yyyy-mm-dd.
get_inmet_data_by_year(2000, first.day = "2000-01-01", last.day = "2000-12-31")
} # }

# This Should get all 2000 data.
get_inmet_data_by_year(2000)
#> # A tibble: 25,176 × 19
#>    station time     precipitation atm_pressure atm_pressure_max atm_pressure_min
#>    <chr>   <chr>            <dbl>        <dbl>            <dbl>            <dbl>
#>  1 A001    2000/05…         -9999        -9999            -9999            -9999
#>  2 A001    2000/05…         -9999        -9999            -9999            -9999
#>  3 A001    2000/05…         -9999        -9999            -9999            -9999
#>  4 A001    2000/05…         -9999        -9999            -9999            -9999
#>  5 A001    2000/05…         -9999        -9999            -9999            -9999
#>  6 A001    2000/05…         -9999        -9999            -9999            -9999
#>  7 A001    2000/05…         -9999        -9999            -9999            -9999
#>  8 A001    2000/05…         -9999        -9999            -9999            -9999
#>  9 A001    2000/05…         -9999        -9999            -9999            -9999
#> 10 A001    2000/05…         -9999        -9999            -9999            -9999
#> # ℹ 25,166 more rows
#> # ℹ 13 more variables: radiation <dbl>, temperature_air <dbl>,
#> #   temperature_dew <dbl>, temperature_max <dbl>, temperature_min <dbl>,
#> #   temperature_dew_max <dbl>, temperature_dew_min <dbl>, humidity_max <dbl>,
#> #   humidity_min <dbl>, humidity <dbl>, wind_direction <dbl>,
#> #   wind_burst_max <dbl>, wind_burst <dbl>